src/findaband_bot/requirements.txt
python-dotenv==0.21.0
spotipy==2.20.0
praw==7.6.0